server/requirements.txt
# Speech Recognition Libraries
speechrecognition==3.10.0
openai-whisper==20231117
vosk==0.3.45

# Audio Processing
PyAudio==0.2.14
wave==0.0.2

# Utilities
numpy>=1.24.0
json5==0.9.14
python-dotenv==1.0.0
